requirements.txt
streamlit
firebase-admin
firebase
pandas
requests
pyrebase4
streamlit-authenticator
python-jwt